[feat] 탈퇴하기 기능
이제 노트를 삭제할 수 있어요. (탈퇴기능)
태그 삭제시 사용자의 id를 참조하지 않는 문제를 해결했어요.

+++ src/app/api/delete-tag/route.ts
@@ -9,14 +9,15 @@ export async function DELETE(request: Request) {
         // Parse the URL or request body for the tagname
         const url = new URL(request.url);
         const tagname = url.searchParams.get('tagname');
+        const userid = url.searchParams.get('userid');
 
-        if (!tagname) {
+        if (!tagname || !userid) {
             return new Response('Missing tagname parameter.', { status: 400 });
         }
 
         // Delete the document with the specified tagname
-        const result = await db.collection<Tag>('tags').deleteOne({ tagname });
-
+        const result = await db.collection<Tag>('tags').deleteOne({ tagname,userid });
+        console.log(result);
         if (result.deletedCount === 0) {
             return new Response('No document found with the specified tagname.', { status: 404 });
         }
